Refactor tetronimoes examples.
