+++ doc/src/mc-tutorial/tetronimoes.md
@@ -17,12 +17,19 @@ and you get something very interesting.
 
 * Objective: Show how to add multiple sites to a body.
 * File: `hoomd-rs/examples/mc-tutorial/tetronimoes.rs`
-* To build and run: `cargo run --release --features "bevy" --example tetronimoes`
+* Run (interactively):
+  ```shell
+  cargo run --release --features "bevy" --example tetronimoes
+  ```
+* Run (in batch mode):
+  ```shell
+  cargo run --release --example tetronimoes
+  ```
 
 ## Use Declarations
 
 ```rust,ignore
-{{#include ../../../examples/mc-tutorial/tetronimoes.rs:use}}
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:use}}
 ```
 
 ## Type Aliases
@@ -31,7 +38,7 @@ Create type aliases for your model's *vector*, *body properties*, and *site
 properties* types so that you don't need to repeat the full nested generic type
 names throughout the code:
 ```rust,ignore
-{{#include ../../../examples/mc-tutorial/tetronimoes.rs:type_aliases}}
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:type_aliases}}
 ```
 The `OrientedPoint` type gives the tetronimo bodies in this tutorial both a
 position in space and an orientation that rotates about the origin of the body.
@@ -40,27 +47,26 @@ site properties.
 
 ## Custom Trial Move
 
-Let's implement a custom trial move to make the tetronimoes move in the way
-you might expect. Like in the [Custom Random Walk], tetronimoes take discrete
-steps left, right, down, or up. Tetronimoes can also rotate by `$ \pm \pi/2 $`.
-Here is the complete code:
+Implement a custom trial move to make the tetronimoes move in the way you might
+expect. Like in the [Custom Random Walk], tetronimoes take discrete steps left,
+right, down, or up. Tetronimoes can also rotate by `$ \pm \pi/2 $`. The
+`DiscreteRotateOrTranslate` type implements this behavior:
 ```rust,ignore
-{{#include ../../../examples/mc-tutorial/tetronimoes.rs:local_trial_all}}
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:local_trial}}
 ```
 
 ### Enumerate Possible Moves
 
-`DiscreteRotateOrTransLate` implements `LocalTrial` by first enumerating the
-possible moves:
+First, enumerate the possible moves:
 ```rust,ignore
-{{#include ../../../examples/mc-tutorial/tetronimoes.rs:local_trial_steps}}
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:local_trial_steps}}
 ```
 
 ### Choose and Propose a Move
 
-Then it chooses a random move and mutates the body properties accordingly:
+Then, choose a random move and mutates the body properties accordingly:
 ```rust,ignore
-{{#include ../../../examples/mc-tutorial/tetronimoes.rs:local_trial_mut}}
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:local_trial_mut}}
 ```
 
 Using `random_bool()`, this code proposes translate moves more often than rotate
@@ -68,48 +74,74 @@ moves because the result is more visually interesting.
 
 ## The Simulation Model
 
+Here is the type that holds the simulation model:
 ```rust,ignore
-{{#include ../../../examples/mc-tutorial/tetronimoes.rs:simulation_new}}
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:simulation_struct}}
 ```
 
-Construct the simulation model as in the [Applying Interactions] tutorial
-with a few differences:
+### Construct the Simulation Model
 
-### Trial Moves
+The `new()` method constructs a new simulation model:
+```rust,ignore
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:simulation_new}}
+```
+
+#### Parameters
+
+Assign all the model parameters in one code block so that they are easy to modify:
+```rust,ignore
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:parameters}}
+```
+
+#### Microstate
+
+Confine the bodies and sites inside of a closed square. Start with no bodies in
+the microstate.
+```rust,ignore
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:microstate}}
+```
+
+#### Hamiltonian
+
+Use the same Hamiltonian as the [Applying Interactions] tutorial:
+```rust,ignore
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:hamiltonian}}
+```
+
+#### Trial Moves
 
 Apply sweeps of the custom `DiscreteRotateOrTranslate` trial move:
 ```rust,ignore
-{{#include ../../../examples/mc-tutorial/tetronimoes.rs:trial_moves}}
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:trial_moves}}
 ```
 
-### Prepare the Tetronimo Shapes
+#### Prepare the Tetronimo Shapes
 
 `new()` also prepares a list of tetronimo shapes for later use. There are
 five types of tetronimoes that are each represented by a vector of
 four points:
 ```rust,ignore
-{{#include ../../../examples/mc-tutorial/tetronimoes.rs:template_sites}}
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:template_sites}}
 ```
 
-### The Tetronimoes Struct
+## Implement `Simulation`
 
-Here is the type that `new()` constructs:
+The `Simulation` implementation closely follows that in [Applying Interactions].
 ```rust,ignore
-{{#include ../../../examples/mc-tutorial/tetronimoes.rs:simulation_struct}}
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:impl_simulation}}
 ```
 
-## Advancing the Simulation
+### Advance the Simulation
 
-To advance the tetronimo simulation forward one step, follow mostly the same
-procedure used in [Applying Interactions]:
 ```rust,ignore
-{{#include ../../../examples/mc-tutorial/tetronimoes.rs:impl_simulation}}
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:advance}}
 ```
-### Adding New Tetronimoes
+
+#### Add New Tetronimoes
 
 The code that adds tetronimoes is more complex than that for disks:
 ```rust,ignore
-{{#include ../../../examples/mc-tutorial/tetronimoes.rs:add}}
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:add}}
 ```
 It first chooses a random tetronimo from the `template_sites`, then it adds the
 body near the top of the boundary with a default orientation of `$ \theta = 0 $`
@@ -123,6 +155,25 @@ use random numbers in your code, you can get a `Rng` to generate them by calling
 > Whenever you use `counter.make_rng`, You *MUST* indicate that your substep is
 > complete by calling `microstate.increment_substep()` so that the next substep
 > will use a different set of random numbers.
+
+#### Apply Trial Moves
+
+Apply the custom trial move to each body in the microstate:
+```rust,ignore
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:apply}}
+```
+
+#### Reset the Simulation
+
+```rust,ignore
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:reset}}
+```
+
+### Get the Simulation Step
+
+```rust,ignore
+{{#rustdoc_include ../../../examples/mc-tutorial/tetronimoes.rs:step}}
+```
 
 ## Conclusion
